Removes all props from CharacterHxForm
